refactor(dashboard): extract QR-modal surface into useQrModal
Move the QR-modal state (linking + per-session share QR, typeable pairing
code #5512, Discord pairing-link delivery #5513) and its two effects
(share-mode reset, pairing-ID auto-refresh #2916) out of App.tsx into a
dedicated hook. Fetch paths, deps arrays, and eslint-disable lines are
byte-identical. The hook is called at handleShowQr's original position so
the Tauri menu wiring still binds the same callback.

Part of #5560 (App.tsx Shell decomposition).
