+++ packages/dashboard/src/App.tsx
@@ -81,6 +81,7 @@ import { useInterventionPing } from './hooks/useInterventionPing'
 import { useShortcutDispatch } from './hooks/useShortcutDispatch'
 import { useChatMessages, toChatViewMessage } from './hooks/useChatMessages'
 import { useTunnelReady } from './hooks/useTunnelReady'
+import { useQrModal } from './hooks/useQrModal'
 import { SplitPane, type SplitDirection } from './components/SplitPane'
 import { persistSidebarWidth, loadPersistedSidebarWidth, persistSplitMode, loadPersistedSplitMode, persistShowConsoleTab, loadPersistedShowConsoleTab, persistInterventionPing, loadPersistedInterventionPing, loadPersistedSidebarPanelHeight, loadPersistedSidebarPanelView, loadPersistedSidebarPanelCollapsed, loadPersistedSidebarRepoOrder, loadPersistedSidebarSessionOrder, persistSidebarRepoOrder, persistSidebarSessionOrder, persistSessionTabOrder, loadPersistedSessionTabOrder } from './store/persistence'
 import { applyOrderById } from './utils/reorderById'
@@ -611,14 +612,6 @@ export function App() {
   const [sessionCreateError, setSessionCreateError] = useState<string | null>(null)
   const [fileAttachments, setFileAttachments] = useState<FileAttachment[]>([])
   const [imageAttachments, setImageAttachments] = useState<ImageAttachment[]>([])
-  const [qrModalOpen, setQrModalOpen] = useState(false)
-  const [qrSvg, setQrSvg] = useState<string | null>(null)
-  const [qrLoading, setQrLoading] = useState(false)
-  const [qrError, setQrError] = useState<string | null>(null)
-  // #5512 — the typeable short pairing code shown beside the linking-mode QR so
-  // camera-less devices can type it instead of scanning. The QR encodes the same
-  // id. Null for per-session "Share" QRs (those carry a session-bound token).
-  const [qrPairingCode, setQrPairingCode] = useState<string | null>(null)
   const [settingsOpen, setSettingsOpen] = useState(() => {
     const params = new URLSearchParams(window.location.search)
     return params.get('settings') === '1'
@@ -1655,75 +1648,22 @@ export function App() {
     setImageAttachments(prev => prev.filter((_, i) => i !== index))
   }, [])
 
-  const fetchQrInto = useCallback(async (path: string) => {
-    setQrModalOpen(true)
-    setQrLoading(true)
-    setQrError(null)
-    setQrSvg(null)
-    setQrPairingCode(null)
-    const token = getAuthToken()
-    if (!token) {
-      setQrLoading(false)
-      setQrError('No auth token available')
-      return
-    }
-    // The typeable code (#5512) only applies to the linking-mode QR — per-session
-    // "Share" QRs (/qr/session/…) issue a session-bound token with no displayed
-    // code. Fetch the code in parallel with the QR for the linking-mode path.
-    const isLinkingQr = path === '/qr'
-    try {
-      const [qrRes, codeRes] = await Promise.all([
-        fetch(path, { headers: { Authorization: `Bearer ${token}` } }),
-        isLinkingQr
-          ? fetch('/pairing-code', { headers: { Authorization: `Bearer ${token}` } }).catch(() => null)
-          : Promise.resolve(null),
-      ])
-      if (!qrRes.ok) {
-        const body = await qrRes.json().catch(() => ({ error: 'Request failed' }))
-        setQrError(body.error || `HTTP ${qrRes.status}`)
-        setQrSvg(null)
-      } else {
-        const svg = await qrRes.text()
-        setQrSvg(svg)
-        setQrError(null)
-      }
-      if (codeRes && codeRes.ok) {
-        const body = await codeRes.json().catch(() => null)
-        if (body?.code) setQrPairingCode(String(body.code))
-      }
-    } catch (err) {
-      setQrError(err instanceof Error ? err.message : 'Failed to fetch QR code')
-      setQrSvg(null)
-    } finally {
-      setQrLoading(false)
-    }
-  }, [])
-
-  const handleShowQr = useCallback(() => fetchQrInto('/qr'), [fetchQrInto])
-
-  // #5513 — host-triggered Discord pairing-link delivery. POSTs to the daemon's
-  // primary-class-gated /pair-discord, which mints a FRESH approval-gated id and
-  // posts only the chroxy:// link. Redeeming it from the channel still needs host
-  // approval, so the channel grants nothing on its own. Returns a result the
-  // QrModal renders inline. Never surfaces token material.
-  const handlePostPairLinkToDiscord = useCallback(async () => {
-    const token = getAuthToken()
-    if (!token) return { posted: false as const, reason: 'no_token' }
-    try {
-      const res = await fetch('/pair-discord', {
-        method: 'POST',
-        headers: { Authorization: `Bearer ${token}`, Accept: 'application/json' },
-      })
-      const body = await res.json().catch(() => null)
-      if (res.ok && body?.posted) {
-        return { posted: true as const, expiresInSeconds: body.expiresInSeconds }
-      }
-      // Auth/availability failures arrive as { error: ... }, not { reason: ... }.
-      return { posted: false as const, reason: body?.reason || body?.error || `http_${res.status}` }
-    } catch {
-      return { posted: false as const, reason: 'post_failed' }
-    }
-  }, [])
+  // #5560 — the QR-modal surface (linking-mode QR, per-session "Share" QR,
+  // typeable pairing code, Discord pairing-link delivery) lives in `useQrModal`.
+  // `handleShowQr` is consumed by the Tauri menu wiring directly below, so the
+  // hook is called here (not lower with the rest of the modal handlers).
+  const {
+    qrModalOpen,
+    setQrModalOpen,
+    qrSvg,
+    qrLoading,
+    qrError,
+    qrPairingCode,
+    qrShareMode,
+    handleShowQr,
+    handleShareSession,
+    handlePostPairLinkToDiscord,
+  } = useQrModal(activeSessionId, pairingRefreshedCount)
 
   // #4695 / #4942 — bridge the macOS menu bar items to App-state
   // handlers. The sidebar's per-project "+" row and the command
@@ -1798,33 +1738,6 @@ export function App() {
     onTunnelSettings: menuOpenSettings,
     onPreferences: menuOpenSettings,
   })
-
-  // #3070: per-session "Share this session" QR. Issues a token bound to the
-  // active session — the scanner can chat into it but cannot list/switch
-  // others. Distinct from the linking-mode QR above, which lets the paired
-  // device manage every session.
-  const [qrShareMode, setQrShareMode] = useState<'link' | 'share'>('link')
-  const handleShareSession = useCallback(() => {
-    if (!activeSessionId) return
-    setQrShareMode('share')
-    void fetchQrInto(`/qr/session/${encodeURIComponent(activeSessionId)}`)
-  }, [activeSessionId, fetchQrInto])
-  // Reset share-mode label whenever the modal reopens via the regular QR
-  // button so the title reflects the actual content.
-  useEffect(() => {
-    if (qrModalOpen && qrShareMode === 'share') return
-    if (!qrModalOpen) setQrShareMode('link')
-  }, [qrModalOpen, qrShareMode])
-
-  // Auto-refresh QR when the server regenerates the pairing ID (#2916).
-  // Only refresh while the modal is open — guarding on qrSvg would reopen
-  // the modal after the user closes it if qrSvg was not cleared on close.
-  useEffect(() => {
-    if (pairingRefreshedCount === 0) return
-    if (!qrModalOpen) return
-    handleShowQr()
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [pairingRefreshedCount])
 
   const handleBannerApprove = useCallback((requestId: string, notificationId: string) => {
     sendPermissionResponse(requestId, 'allow')
